Adding templates (not in use, just the CRUD APIs and no document revisions).

+++ src/main/java/org/sagebionetworks/bridge/hibernate/HibernateTemplateDao.java
@@ -3,6 +3,8 @@ package org.sagebionetworks.bridge.hibernate;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static org.sagebionetworks.bridge.models.ResourceList.OFFSET_BY;
 import static org.sagebionetworks.bridge.models.ResourceList.PAGE_SIZE;
+import static org.sagebionetworks.bridge.models.ResourceList.TEMPLATE_TYPE;
+import static org.sagebionetworks.bridge.models.ResourceList.TOTAL;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,15 +13,28 @@ import javax.annotation.Resource;
 
 import com.google.common.collect.ImmutableMap;
 
+import org.springframework.stereotype.Component;
+
 import org.sagebionetworks.bridge.dao.TemplateDao;
 import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.Template;
 import org.sagebionetworks.bridge.models.TemplateType;
 import org.sagebionetworks.bridge.models.studies.StudyIdentifier;
 
+/**
+ * DAO implementation for templates. All the business constraints are enforced in TemplateService because they are not 
+ * specific to our persistence store.
+ */
+@Component
 public class HibernateTemplateDao implements TemplateDao {
-    private static final String GET_QUERY = "SELECT template FROM HibernateTemplate as template " + 
+    private static final String SELECT_TEMPLATE = "SELECT template ";
+    private static final String SELECT_COUNT = "SELECT count(*) ";
+    
+    private static final String GET_ALL = "FROM HibernateTemplate as template " + 
             "WHERE templateType = :templateType AND studyId = :studyId ORDER BY createdOn DESC";
+    
+    private static final String GET_ACTIVE = "FROM HibernateTemplate as template " + 
+            "WHERE templateType = :templateType AND studyId = :studyId AND deleted = 0 ORDER BY createdOn DESC";
     
     private HibernateHelper hibernateHelper;
     
@@ -29,31 +44,43 @@ public class HibernateTemplateDao implements TemplateDao {
     }
     
     @Override
-    public PagedResourceList<? extends Template> getTemplates(StudyIdentifier studyId, TemplateType type, int offset, int pageSize) {
+    public PagedResourceList<? extends Template> getTemplates(StudyIdentifier studyId, TemplateType type,
+            Integer offset, Integer pageSize, boolean includeDeleted) {
+        checkNotNull(studyId);
         checkNotNull(type);
         
-        ImmutableMap<String,Object> params = ImmutableMap.of("templateType", type.name());
-
-        int count = hibernateHelper.queryCount(GET_QUERY, params);
-        List<? extends HibernateTemplate> results = hibernateHelper.queryGet(GET_QUERY, params, offset, pageSize, HibernateTemplate.class);
+        ImmutableMap<String,Object> params = ImmutableMap.of(
+                TEMPLATE_TYPE, type, 
+                "studyId", studyId.getIdentifier());
+        String countQuery = SELECT_COUNT + ((!includeDeleted) ? GET_ACTIVE : GET_ALL);
+        String getQuery = SELECT_TEMPLATE + ((!includeDeleted) ? GET_ACTIVE : GET_ALL);
         
-        return new PagedResourceList<>(results, count)
+        int total = hibernateHelper.queryCount(countQuery, params);
+        List<? extends HibernateTemplate> results = hibernateHelper.queryGet(
+                getQuery, params, offset, pageSize, HibernateTemplate.class);
+        
+        return new PagedResourceList<>(results, total)
+                .withRequestParam(TEMPLATE_TYPE, type)
                 .withRequestParam(PAGE_SIZE, pageSize)
-                .withRequestParam(OFFSET_BY, offset);
+                .withRequestParam(OFFSET_BY, offset)
+                .withRequestParam(TOTAL, total);
     }
     
     @Override
-    public Optional<Template> getTemplate(String guid) { 
+    public Optional<Template> getTemplate(StudyIdentifier studyId, String guid) { 
         checkNotNull(guid);
         
         Template template = hibernateHelper.getById(HibernateTemplate.class, guid);
+        if (template != null && !template.getStudyId().equals(studyId.getIdentifier())) {
+            return Optional.empty();
+        }
         return Optional.ofNullable(template);
     }
 
     @Override
     public void createTemplate(Template template) {
         checkNotNull(template);
-
+        
         hibernateHelper.create(template, null); 
     }
 
@@ -65,10 +92,13 @@ public class HibernateTemplateDao implements TemplateDao {
     }
 
     @Override
-    public void deleteTemplatePermanently(String guid) {
+    public void deleteTemplatePermanently(StudyIdentifier studyId, String guid) {
         checkNotNull(guid);
-
-        hibernateHelper.deleteById(HibernateTemplate.class, guid);
+        
+        Optional<Template> template = getTemplate(studyId, guid);
+        if (template.isPresent()) {
+            hibernateHelper.deleteById(HibernateTemplate.class, guid);    
+        }
     }
 
 }
